pyproject.toml
[project]
name = "ragql"
version = "0.1.0"
description = "RAG-based Text-to-SQL conversion system using Mistral AI and Neon PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-mistralai>=0.1.0",
    "langchain-postgres>=0.0.6",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "psycopg[binary]>=3.2.9",
    "transformers>=4.55.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0"
]

[project.scripts]
ragql = "ragql.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
